Update processLogMavens.py
Add command line option to specify a specific player or players to restrict the output to.  Can be specified multiple times to get more than one player. Indicates if no output found for player.

Modifies command line output and email recipients if that option is enabled. Does not restrict CSV output if that option is enabled.